+++ ThirdPartyAdapters/chartboost/chartboost/src/test/kotlin/com/google/ads/mediation/chartboost/ChartboostMediationAdapterTest.kt
@@ -16,6 +16,7 @@ import com.google.ads.mediation.chartboost.ChartboostConstants.ERROR_INVALID_SER
 import com.google.ads.mediation.chartboost.ChartboostInitializer.getInstance
 import com.google.ads.mediation.chartboost.ChartboostMediationAdapter.ERROR_MESSAGE_INVALID_SERVER_PARAMETERS
 import com.google.ads.mediation.chartboost.ChartboostMediationAdapter.ERROR_MESSAGE_MISSING_OR_INVALID_APP_ID
+import com.google.android.gms.ads.AgeRestrictedTreatment
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
@@ -53,6 +54,7 @@ class ChartboostMediationAdapterTest {
           RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
         )
         .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .setAgeRestrictedTreatment(AgeRestrictedTreatment.UNSPECIFIED)
         .build()
     )
     ChartboostInitializer.clearInstance()
@@ -353,7 +355,27 @@ class ChartboostMediationAdapterTest {
       adapter.initialize(context, initializationCompleteCallback, listOf(mediationConfiguration))
 
       it.verify { Chartboost.addDataUseConsent(any(), coppaCaptor.capture()) }
-      assertThat(coppaCaptor.firstValue.consent).isFalse()
+    }
+  }
+
+  @Test
+  fun initialize_withAgeRestrictedTreatmentChild_updatesCoppaTrue() {
+    MobileAds.setRequestConfiguration(
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .setAgeRestrictedTreatment(AgeRestrictedTreatment.CHILD)
+        .build()
+    )
+    val coppaCaptor = argumentCaptor<COPPA>()
+    ChartboostMediationAdapter.setAppParams("app_id", "app_signature")
+    mockStatic(Chartboost::class.java).use {
+      adapter.initialize(context, initializationCompleteCallback, listOf(mediationConfiguration))
+
+      it.verify { Chartboost.addDataUseConsent(any(), coppaCaptor.capture()) }
+      assertThat(coppaCaptor.firstValue.consent).isTrue()
     }
   }
 }